[refactor] componentization
